compact metrics

+++ cmd/ingest_logs_cmd.go
@@ -172,7 +172,7 @@ func (wm *writerManager) getWriter(key hourSlotKey) (*parquetwriter.UnifiedWrite
 
 	// Create new writer
 	baseName := fmt.Sprintf("logs_%s_%d_%d_%d", wm.orgID, key.dateint, key.hour, key.slot)
-	writer, err := factories.NewLogsWriter(baseName, wm.tmpdir, 50*1024*1024, float64(wm.rpfEstimate))
+	writer, err := factories.NewLogsWriter(baseName, wm.tmpdir, 50*1024*1024, wm.rpfEstimate)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create logs writer: %w", err)
 	}
